Initial Gutenberg compatability
This updates the meta box to refresh the meta table when a Gutenberg post is saved.  It does this very simply by just re-requesting the table HTML, rather than getting fancy with something like a JSON object and parsing it into DOM bits.  Not elegant, but simple.

Might fix #17.  Thoughts @danielbachhuber?